+++ RTextNpp/ViewModels/LinkTargetModel.cs
@@ -47,12 +47,12 @@ namespace RTextNppPlugin.ViewModels
             _targets.AddRange(aLinkTargetModels);
         }
 
-        internal void OnZoomLevelChanged(int newZoomLevel)
+        internal void OnZoomLevelChanged(double newZoomLevel)
         {
             //calculate actual zoom level , based on Scintilla zoom factors...
 
             //try 8% increments / decrements
-            ZoomLevel = (1 + (Constants.ZOOM_FACTOR * newZoomLevel));
+            ZoomLevel = (1.0 + (Constants.ZOOM_FACTOR * newZoomLevel));
         }
 
         public double ZoomLevel
